Fixed Bugs with routes - Fixed Verticals routes with Create, Get, Delete - Added Containers with Data and Descriptor for nodes - Updated schemas and pydantic models for nodes - Created Sensor_types routes with Create,Get,Delete but needed to check!!!

+++ app/routes/verticals.py
@@ -9,7 +9,8 @@ from app.utils.om2m import Om2m
 from app.schemas.verticals import VerticalCreate, VerticalGetAll, VerticalDelete
 import xml.etree.ElementTree as ET
 from app.models.vertical import Vertical as DBAE
-
+import json
+from app.utils.utils import gen_vertical_code
 
 router = APIRouter()
 
@@ -20,7 +21,10 @@ om2m = Om2m("admin", "admin", "http://localhost:8080/~/in-cse/in-name")
 @token_required
 @admin_required
 def create_ae(
-    vertical: VerticalCreate, request: Request, session: Session = Depends(get_session)
+    vertical: VerticalCreate,
+    request: Request,
+    session: Session = Depends(get_session),
+    current_user=None,
 ):
     """
     Create an AE (Application Entity) with the given name and labels.
@@ -37,11 +41,18 @@ def create_ae(
         HTTPException: If there is an error creating the AE or if the AE already exists.
     """
     ae_name = vertical.ae_name
-    status_code, data = om2m.create_ae(ae_name, vertical.path, labels=[ae_name])
+    assigned_name = gen_vertical_code(vertical.ae_name)
+    if len(vertical.labels) == 0:
+        labels = [ae_name]
+    status_code, data = om2m.create_ae(assigned_name, vertical.path, labels=labels)
     if status_code == 201:
-        new_ae = DBAE(ae_name=ae_name, path=vertical.path)
-        session.add(new_ae)
+        res_id = json.loads(data)["m2m:ae"]["ri"].split("/")[-1]
+        # TODO: Add the AE to the database
+        db_vertical = DBAE(res_name=assigned_name, labels=labels, orid=res_id)
+        session.add(db_vertical)
         session.commit()
+        print("AE created")
+        raise HTTPException(status_code=201, detail="AE created")
     elif status_code == 409:
         raise HTTPException(status_code=409, detail="AE already exists")
     else:
@@ -49,8 +60,7 @@ def create_ae(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail="Error creating AE",
         )
-
-    return status_code
+    # return status_code
 
 
 @router.get("/get-aes")
@@ -72,7 +82,7 @@ def get_aes(
 
     Returns:
     - list: A list of dictionaries containing the "rn" and "ri" attributes of each subcontainer.
-    
+
     Raises:
     - HTTPException: If the path is not found or there is an error parsing XML.
     - Exception: If there is an error retrieving AE.
@@ -117,7 +127,10 @@ def get_aes(
 @token_required
 @admin_required
 def delete_ae(
-    vertical: VerticalDelete, request: Request, session: Session = Depends(get_session)
+    vertical: VerticalDelete,
+    request: Request,
+    session: Session = Depends(get_session),
+    current_user=None,
 ):
     """
     This function deletes an AE resource in OM2M.
@@ -130,16 +143,21 @@ def delete_ae(
     Returns:
         int: The status code of the request.
     """
-
     ae = session.query(DBAE).filter(DBAE.res_name == vertical.ae_name).first()
     if ae is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="AE not found"
         )
-    status_code = om2m.delete_resource(f"{vertical.ae_name}")
-    if status_code == 200:
+    if vertical.path == "":
+        final_path = f"{vertical.ae_name}"
+    else:
+        final_path = f"{vertical.path}/{vertical.ae_name}"
+    status_code = om2m.delete_resource(final_path).status_code
+    print(status_code)
+    if status_code >= 200 and status_code < 300:
         session.delete(ae)
         session.commit()
+        raise HTTPException(status_code=200, detail="AE deleted")
     elif status_code == 404:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="AE not found"
